Bug 1939704: add axios proxy config (#1172)
* add axios proxy config

* fix arguments

* cleanup
