+++ Dockerfile
@@ -10,7 +10,6 @@ COPY --from=builder /mig-ui/dist /opt/app-root/src/staticroot
 COPY --from=builder /mig-ui/public/favicon.ico /opt/app-root/src/staticroot
 COPY --from=builder /mig-ui/public/index.ejs /opt/app-root/src/staticroot
 COPY --from=builder /mig-ui/deploy/main.js /opt/app-root/src
-COPY --from=builder /mig-ui/deploy/oAuthHelpers.js /opt/app-root/src
 COPY --from=builder /mig-ui/node_modules /opt/app-root/src/node_modules
 
 ENV MIGMETA_FILE="/etc/mig-ui/migmeta.json"
